update : migrate ErrorResponse schema to exception module to handle the circular import error
