+++ tool/backend/src/core/__init__.py
@@ -1,5 +1,25 @@
 from .configs import settings
+from .exceptions import (
+    BaseAPIException,
+    BadRequestException,
+    ConflictException,
+    UnauthorizedException,
+    ForbiddenException,
+    NotFoundException,
+    UnprocessableEntityException,
+    InternalServerException,
+    api_exception_handler
+)
 
 __all__ = [
-    "settings"
+    "settings",
+    "BaseAPIException",
+    "BadRequestException",
+    "ConflictException",
+    "UnauthorizedException",
+    "ForbiddenException",
+    "NotFoundException",
+    "UnprocessableEntityException",
+    "InternalServerException",
+    "api_exception_handler"
 ]
